refactor(dashboard/files): ditch hooks, controllers, stores, lib folders and squash them into shared separated by domain
