+++ apps/dashboard/src/app/[lang]/(auth)/onboarding/page.tsx
@@ -17,9 +17,9 @@ import {
 } from "@/components/auth/onboarding-sign-up-form";
 import { StoreConfigurationOnboardingForm } from "@/components/auth/onboarding-store-configuration-form";
 import { StoreSetupOnboardingForm } from "@/components/auth/onboarding-store-setup-form";
-import { useOnboardingController } from "@/hooks/controllers/use-onboarding-controller";
 import { authClient } from "@/shared/api/auth-client";
 import { RoutePaths } from "@/shared/config/routes";
+import { useOnboardingController } from "@/shared/lib/onboarding/controller.hook";
 
 export default function OnboardingPage() {
   const t = useTranslations("onboarding");
