+++ apps/dashboard/src/app/[lang]/(dashboard)/products/new/page.tsx
@@ -12,7 +12,7 @@ import {
   type ProductFormHandle,
 } from "@/components/app/products/product-form";
 import { getRouteWithQuery, RoutePaths } from "@/shared/config/routes";
-import { useActiveStoreStore } from "@/stores/active-store.store";
+import { useActiveStoreStore } from "@/shared/lib/store/active.store";
 
 export default function NewProductPage() {
   const t = useTranslations("products.create");
