+++ apps/dashboard/src/components/app/products/product-form-essentials.tsx
@@ -3,8 +3,8 @@
 import type { SelectOptionGroup } from "@dukkani/ui/components/forms/select-field";
 import { withForm } from "@dukkani/ui/hooks/use-app-form";
 import { useTranslations } from "next-intl";
-import { useCurrentStoreCurrency } from "@/stores";
-import { productFormOptions } from "@/lib/product-form-options";
+import { productFormOptions } from "@/shared/lib/product/form";
+import { useCurrentStoreCurrency } from "@/shared/lib/store/current-currency.hook";
 
 export const ProductFormEssentials = withForm({
   ...productFormOptions,
