+++ apps/dashboard/src/components/app/products/product-form-variants.tsx
@@ -15,7 +15,7 @@ import { Icons } from "@dukkani/ui/components/icons";
 import { withForm } from "@dukkani/ui/hooks/use-app-form";
 import { useTranslations } from "next-intl";
 import { useCallback } from "react";
-import { productFormOptions } from "@/lib/product-form-options";
+import { productFormOptions } from "@/shared/lib/product/form";
 
 export const ProductFormVariants = withForm({
   ...productFormOptions,
