+++ apps/dashboard/src/components/app/products/products-filter-drawer.tsx
@@ -18,7 +18,7 @@ import { Field, FieldLabel } from "@dukkani/ui/components/field";
 import { Input } from "@dukkani/ui/components/input";
 import { useTranslations } from "next-intl";
 import { useEffect, useState } from "react";
-import type { StockFilter, VariantsFilter } from "@/stores/product.store";
+import type { StockFilter, VariantsFilter } from "@/shared/lib/product/store";
 
 interface ProductsFilterDrawerProps {
   open: boolean;
