--- apps/dashboard/src/hooks/use-onboarding.ts
@@ -1,2 +1,0 @@
-import { useQuery } from "@tanstack/react-query";
-import { client } from "@/shared/api/orpc";
