+++ apps/dashboard/src/shared/lib/onboarding/stepper.util.ts
