+++ apps/dashboard/src/shared/lib/order/datetime.util.ts
@@ -1,5 +1,3 @@
-import type { OrderIncludeOutput } from "@dukkani/common/schemas/order/output";
-
 export function formatOrderDateTime(
   date: Date,
   now: Date,
@@ -30,17 +28,4 @@ export function formatOrderDateTime(
     hour: "2-digit",
     minute: "2-digit",
   });
-}
-
-export function getOrderTotal(order: OrderIncludeOutput): number {
-  return (
-    order.orderItems?.reduce(
-      (sum, item) => sum + item.price * item.quantity,
-      0,
-    ) ?? 0
-  );
-}
-
-export function getItemsCount(order: OrderIncludeOutput): number {
-  return order.orderItems?.reduce((sum, item) => sum + item.quantity, 0) ?? 0;
 }
