--- apps/dashboard/src/stores/index.ts
@@ -1,7 +1,0 @@
-export {
-  useActiveStoreStore,
-  useCurrentStoreCurrency,
-  useFormatPriceForActiveStore,
-} from "./active-store.store";
-export { useOrderStore } from "./order.store";
-export { useProductStore } from "./product.store";
